Initial PoC to refactor loopback-boot
