+++ lib/bundler.js
@@ -14,39 +14,37 @@ var g = require('strong-globalize')();
  * @param {Object} instructions Boot instructions.
  * @param {Object} bundler A browserify object created by `browserify()`.
  */
-
-module.exports = function addInstructionsToBrowserify(instructions, bundler) {
-  bundleModelScripts(instructions, bundler);
-  bundleMixinScripts(instructions, bundler);
-  bundleComponentScripts(instructions, bundler);
-  bundleOtherScripts(instructions, bundler);
-  bundleInstructions(instructions, bundler);
+module.exports = function addInstructionsToBrowserify(context, bundler) {
+  bundleModelScripts(context, bundler);
+  bundleMixinScripts(context, bundler);
+  bundleComponentScripts(context, bundler);
+  bundleOtherScripts(context, bundler);
+  bundleInstructions(context, bundler);
 };
 
-function bundleOtherScripts(instructions, bundler) {
-  for (var key in instructions.files) {
-    addScriptsToBundle(key, instructions.files[key], bundler);
-  }
+function bundleOtherScripts(context, bundler) {
+  var list = context.instructions.bootScripts;
+  addScriptsToBundle('boot', list, bundler);
 }
 
-function bundleModelScripts(instructions, bundler) {
-  bundleSourceFiles(instructions, 'models', bundler);
+function bundleModelScripts(context, bundler) {
+  bundleSourceFiles(context, 'models', bundler);
 }
 
-function bundleMixinScripts(instructions, bundler) {
-  bundleSourceFiles(instructions, 'mixins', bundler);
+function bundleMixinScripts(context, bundler) {
+  bundleSourceFiles(context, 'mixins', bundler);
 }
 
-function bundleComponentScripts(instructions, bundler) {
-  bundleSourceFiles(instructions, 'components', bundler);
+function bundleComponentScripts(context, bundler) {
+  bundleSourceFiles(context, 'components', bundler);
 }
 
-function bundleSourceFiles(instructions, type, bundler) {
-  var files = instructions[type]
+function bundleSourceFiles(context, type, bundler) {
+  var files = context.instructions[type]
     .map(function(m) { return m.sourceFile; })
     .filter(function(f) { return !!f; });
 
-  var instructionToFileMapping = instructions[type]
+  var instructionToFileMapping = context.instructions[type]
     .map(function(m) { return files.indexOf(m.sourceFile); });
 
   addScriptsToBundle(type, files, bundler);
@@ -54,7 +52,7 @@ function bundleSourceFiles(instructions, type, bundler) {
   // Update `sourceFile` properties with the new paths
   instructionToFileMapping.forEach(function(fileIx, sourceIx) {
     if (fileIx === -1) return;
-    instructions[type][sourceIx].sourceFile = files[fileIx];
+    context.instructions[type][sourceIx].sourceFile = files[fileIx];
   });
 }
 
@@ -74,14 +72,14 @@ function addScriptsToBundle(name, list, bundler) {
     // Add the file to the bundle.
     bundler.require(filepath, { expose: fileid });
 
-    // Rewrite the instructions entry with the new id that will be
+    // Rewrite the context entry with the new id that will be
     // used to load the file via `require(fileid)`.
     list[ix] = fileid;
   }
 }
 
-function bundleInstructions(instructions, bundler) {
-  instructions = cloneDeep(instructions);
+function bundleInstructions(context, bundler) {
+  var instructions = cloneDeep(context.instructions);
 
   var hasMiddleware = instructions.middleware.phases.length ||
     instructions.middleware.middleware.length;
