+++ test/helpers/browser.js
@@ -58,6 +58,8 @@ function createContext() {
         error: [],
       },
     },
+
+    ArrayBuffer: ArrayBuffer,
   };
 
   // `window` is used by loopback to detect browser runtime
